# greenhouse_emulator/api/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/greenhouses_info/{greenhouse_id}":{"get":{"tags":["greenhouses, meterings"],"summary":"Get Greenhouse","operationId":"get_greenhouse_greenhouses_info__greenhouse_id__get","parameters":[{"name":"greenhouse_id","in":"path","required":true,"schema":{"type":"integer","title":"Greenhouse Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/temperature":{"post":{"tags":["greenhouses, meterings"],"summary":"Get Temperature","operationId":"get_temperature_temperature_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeteringsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GreenhouseMeterings"},"type":"array","title":"Response 200 Get Temperature Temperature Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/humidity":{"post":{"tags":["greenhouses, meterings"],"summary":"Get Humidity","operationId":"get_humidity_humidity_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeteringsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GreenhouseMeterings"},"type":"array","title":"Response 200 Get Humidity Humidity Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ph":{"post":{"tags":["greenhouses, meterings"],"summary":"Get Ph","operationId":"get_ph_ph_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeteringsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GreenhouseMeterings"},"type":"array","title":"Response 200 Get Ph Ph Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GreenhouseMeterings":{"properties":{"id":{"type":"integer","title":"Id"},"data":{"$ref":"#/components/schemas/MeteringPoint"}},"type":"object","required":["id","data"],"title":"GreenhouseMeterings"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MeteringPoint":{"items":{"prefixItems":[{"type":"string","format":"date-time"},{"anyOf":[{"type":"number"},{"type":"null"}]}],"type":"array","maxItems":2,"minItems":2},"type":"array","title":"MeteringPoint"},"MeteringsRequest":{"properties":{"greenhouses":{"items":{"type":"integer"},"type":"array","title":"Greenhouses"},"dt_from":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Dt From"},"dt_to":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Dt To"}},"type":"object","required":["greenhouses"],"title":"MeteringsRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}