# greenhouse_emulator/src/api_server/app/greenhouses/schema/responses/success_responses.py
from datetime import datetime
from typing import List, Optional, Tuple

from pydantic import BaseModel, RootModel


class MeteringPoint(RootModel[List[Tuple[datetime, Optional[float]]]]):
    pass


class GreenhouseMeterings(BaseModel):
    id: int
    data: MeteringPoint


GreenhousesMeteringsResponse = list[GreenhouseMeterings]